Better separation in e2e tests
