+++ backend/tests/e2e/main.rs
@@ -2,5 +2,6 @@
 // This file allows cargo test --test e2e to run all E2E tests
 // E2E tests are ignored by default in CI - run with: cargo test --test e2e -- --ignored
 
+mod test_setup;
 mod upload_e2e;
 mod utils;
