+++ backend/tests/e2e/utils.rs
@@ -4,8 +4,6 @@ use hex;
 use rand::RngCore;
 use reqwest::header::{HeaderMap, HeaderValue, CONTENT_LENGTH, CONTENT_TYPE};
 use sha2::{Digest, Sha256};
-use std::time::Duration;
-use tokio::time::sleep;
 
 /// Generate test image data with specified size and return data + SHA-256 hash
 pub fn generate_test_image(size: usize) -> (Vec<u8>, String) {
@@ -87,34 +85,9 @@ pub async fn s3_object_exists(
     }
 }
 
-/// Wait for specified duration (useful for expiration tests)
-pub async fn wait_for_duration(duration: Duration) {
-    sleep(duration).await;
-}
-
-/// Create test data with specific content that will produce a known SHA-256
-pub fn create_test_data_with_known_hash() -> (Vec<u8>, String) {
-    let data = b"Hello, World! This is test data for E2E upload testing.";
-    let sha256 = calculate_sha256(data);
-    (data.to_vec(), sha256)
-}
-
-/// Create test data with wrong checksum (for bad actor tests)
-pub fn create_data_with_wrong_checksum() -> (Vec<u8>, String, String) {
-    let data = b"This is the actual data";
-    let wrong_sha256 = "a".repeat(64); // Invalid SHA-256 that will fail validation
-    let correct_sha256 = calculate_sha256(data);
-    (data.to_vec(), wrong_sha256, correct_sha256)
-}
-
 /// Verify that two byte arrays are identical
 pub fn verify_data_integrity(original: &[u8], downloaded: &[u8]) -> bool {
     original == downloaded
-}
-
-/// Extract S3 key from asset_id returned by the API
-pub fn extract_s3_key_from_asset_id(asset_id: &str) -> String {
-    asset_id.to_string()
 }
 
 /// Create headers for S3 upload with specific content type and checksum
@@ -139,10 +112,4 @@ pub fn create_upload_headers(
     }
 
     headers
-}
-
-/// Parse presigned URL to extract query parameters (useful for debugging)
-pub fn parse_presigned_url_params(url: &str) -> std::collections::HashMap<String, String> {
-    let parsed = url::Url::parse(url).unwrap();
-    parsed.query_pairs().into_owned().collect()
 }
